Take temporary attribute into account and save normalized coordinates instead of actual ones

+++ database/migrations/2017_03_15_143456_create_field_areas_table.php
@@ -23,10 +23,10 @@ class CreateFieldAreasTable extends Migration
         Schema::create('field_areas', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('page')->unsigned();
-            $table->integer('x')->unsigned();
-            $table->integer('y')->unsigned();
-            $table->integer('w')->unsigned();
-            $table->integer('h')->unsigned();
+            $table->double('x')->unsigned();
+            $table->double('y')->unsigned();
+            $table->double('w')->unsigned();
+            $table->double('h')->unsigned();
             $table->timestamps();
         });
     }
